requirements.txt
asgiref==3.3.4
Django==2.2.20
django-bootstrap-modal-forms==1.4.1
django-widget-tweaks==1.4.2
mysqlclient==2.0.3
pytz==2018.5
selenium==3.14.0
sqlparse==0.4.1
urllib3==1.26.5

